angular_dependency_ai_agent_python/requirements.txt
openai>=1.0.0
requests
python-dotenv
packaging